+++ indigo/indigo/src/main/scala/indigo/platform/renderer/webgl2/RendererWebGL2.scala
@@ -405,12 +405,12 @@ final class RendererWebGL2(
       lastHeight = actualHeight
 
       orthographicProjectionMatrix =
-        CheapMatrix4.orthographic(actualWidth.toDouble / magnification, actualHeight.toDouble / magnification)
+        CheapMatrix4.orthographic(actualWidth.toFloat / magnification, actualHeight.toFloat / magnification)
       defaultLayerProjectionMatrix = orthographicProjectionMatrix.scale(1.0, -1.0, 1.0).mat.map(_.toFloat)
       orthographicProjectionMatrixNoMag =
-        CheapMatrix4.orthographic(actualWidth.toDouble, actualHeight.toDouble).mat.map(_.toFloat)
+        CheapMatrix4.orthographic(actualWidth.toFloat, actualHeight.toFloat).mat.map(_.toFloat)
       orthographicProjectionMatrixNoMagFlipped =
-        CheapMatrix4.orthographic(actualWidth.toDouble, actualHeight.toDouble).scale(1.0, -1.0, 1.0).mat.map(_.toFloat)
+        CheapMatrix4.orthographic(actualWidth.toFloat, actualHeight.toFloat).scale(1.0, -1.0, 1.0).mat.map(_.toFloat)
 
       layerEntityFrameBuffer = FrameBufferFunctions.createFrameBufferSingle(gl, actualWidth, actualHeight)
       scalingFrameBuffer = FrameBufferFunctions.createFrameBufferSingle(gl, actualWidth, actualHeight)
